moved form into MessagePost component, created component for message text
