+++ code/src/MessagePost.js
@@ -1,25 +1,48 @@
-import React, { useState } from "react"
+import React from "react"
 import "./messagepost.css"
+import { Loader } from "./Loader"
 
 export const MessagePost = (props) => {
-    const [textLength, setTextLength] = useState(0)
+    const { newThought, setNewThought, sendingThought, setSendingThought, setMessages } = props
     const handleEvent = (event) => {
-        props.setNewThought(event.target.value)
-        setTextLength(event.target.value.length)
-
+        setNewThought(event.target.value)
     }
 
-
+    const handleSubmit = (event) => {
+        event.preventDefault();
+        setSendingThought(true)
+        fetch("https://technigo-thoughts.herokuapp.com/", {
+            method: 'POST',
+            headers: { "content-type": "application/json" },
+            body: JSON.stringify({ message: newThought })
+        })
+            .then(res => res.json())
+            .then(latestThought => {
+                setMessages(messages => [latestThought, ...messages])
+                setNewThought('')
+                setSendingThought(false)
+            })
+    }
     return (
-        <div>
-            <p className="post-title">What's making you happy right now?</p>
-            <textarea maxLength={140}
-                className="post-text"
-                onChange={handleEvent} type="text"></textarea>
-            <p className="character-count"
-                style={{ color: textLength >= 100 ? "red" : textLength >= 50 ? "orange" : "black" }}>
-                {140 - textLength} characters left</p>
-            <button className="post-button" disabled={!props.newThought} type="submit" >❤️Send Happy Thought❤️</button>
-        </div>
+        <form className="thoughts-container form" onSubmit={handleSubmit}>
+            {sendingThought && <Loader />}
+            {!sendingThought &&
+                <div>
+                    <p className="post-title">What's making you happy right now?</p>
+                    <textarea maxLength={140}
+                        value={newThought}
+                        className="post-text"
+                        onChange={handleEvent} type="text">
+                    </textarea>
+                    <p className="character-count"
+                        style={{ color: newThought.length >= 100 ? "red" : newThought.length >= 50 ? "orange" : "black" }}>
+                        {140 - newThought.length} characters left
+                </p>
+                    <button className="post-button" disabled={!newThought || newThought.length < 5} type="submit" >
+                        <span role="img" aria-label="heart">❤️</span>
+                        Send Happy Thought<span role="img" aria-label="heart">❤️</span>
+                    </button>
+                </div>}
+        </form >
     )
 }
