+++ code/src/MessageText.js
@@ -1,0 +1,6 @@
+import React from "react"
+
+
+export const MessageText = (props) => {
+    return <p>{props.text}</p>
+}
