+++ code/src/app.css
@@ -18,6 +18,7 @@
     display: flex;
     flex-direction: column;
     padding: 15px;
+    overflow-wrap: break-word;
 }
 
 .likes-and-time {
@@ -27,10 +28,6 @@
     align-items: flex-end;
     color: grey;
     font: .8em helvetica;
-}
-
-.form {
-    background-color: #f2f0f0;
 }
 
 @media (min-width: 680px) {
