+++ code/src/loader.css
@@ -9,7 +9,6 @@
     justify-self: center;
 }
 
-
 @keyframes spin {
     0% {transform: rotate(0deg)}
     100% {transform: rotate(360deg)}
