+++ code/src/messagepost.css
@@ -1,3 +1,7 @@
+.form {
+    background-color: #f2f0f0;
+}
+
 .post-button {
     color: black;
     background-color: #ffadad;
